♻️ [Refactoring]: #455 Phase 12 decode_octal 単体テストを tests/decode_octal.rs に分離（最終整形）
純関数 decode_octal の単体テスト 6 件を
literal_string/tests/decode_octal.rs に切り出し、tests.rs ハブを
`mod {cat};` 12 行のみの最小宣言型に整形（use 文・ヘルパーゼロ）。
全 12 サブファイル < 800 行、関数名集合 73 件で Phase 0 と完全一致。
